requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.13
requests==2.31.0
